add ZoneButton cmp
